finished implementing this project
There's a few issues; I forgot to close files in fput and fget, but I need to call this project done.  Also some issues with the cd command, could use looking into, but this was optional in the first place and was more for my sanity/learning so not worrying about it.